Moved getRenderedHref to testHelpers, switched Discover Links integration tests to TS, and added ML link integration test
